+++ src/spring-cloud/azext_spring_cloud/_params.py
@@ -14,7 +14,8 @@ from ._validators import (validate_env, validate_cosmos_type, validate_resource_
                           validate_tracing_parameters_asc_create, validate_tracing_parameters_asc_update,
                           validate_app_insights_parameters, validate_instance_count, validate_java_agent_parameters,
                           validate_jar)
-from ._validators_enterprise import (only_support_enterprise, validate_builder_resource, validate_builder_create,
+from ._validators_enterprise import (only_support_enterprise, not_support_enterprise,
+                                     validate_builder_resource, validate_builder_create,
                                      validate_builder_update, validate_build_pool_size,
                                      validate_git_uri, validate_acs_patterns, validate_config_file_patterns,
                                      validate_routes, validate_gateway_instance_count,
@@ -268,7 +269,7 @@ def load_arguments(self, _):
     for scope in ['spring-cloud app update', 'spring-cloud app deployment create', 'spring-cloud app deploy', 'spring-cloud app create']:
         with self.argument_context(scope) as c:
             c.argument('runtime_version', arg_type=get_enum_type(SupportedRuntimeValue),
-                       help='Runtime version of used language')
+                       help='Runtime version of used language', validator=not_support_enterprise)
             c.argument('jvm_options', type=str, validator=validate_jvm_options,
                        help="A string containing jvm options, use '=' instead of ' ' for this argument to avoid bash parse error, eg: --jvm-options='-Xms1024m -Xmx2048m'")
             c.argument('env', env_type)
@@ -303,7 +304,7 @@ def load_arguments(self, _):
                 'main_entry', options_list=[
                     '--main-entry', '-m'], help="A string containing the path to the .NET executable relative to zip root.")
             c.argument(
-                'target_module', help='Child module to be deployed, required for multiple jar packages built from source code.', arg_group='Source Code deploy')
+                'target_module', help='Child module to be deployed, required for multiple jar packages built from source code.', arg_group='Source Code deploy', validate=not_support_enterprise)
             c.argument(
                 'version', help='Deployment version, keep unchanged if not set.')
             c.argument(
@@ -318,6 +319,8 @@ def load_arguments(self, _):
                 'container_command', help='The command of the container image.', nargs='*', arg_group='Custom Container')
             c.argument(
                 'container_args', help='The arguments of the container image.', nargs='*', arg_group='Custom Container')
+            c.argument(
+                'build_env', help='The key-value pairs of env used in build phase.', type=dict, validate=only_support_enterprise)
 
     with self.argument_context('spring-cloud app deploy') as c:
         c.argument('source_path', arg_type=source_path_type, validator=validate_deloy_path)