+++ src/spring-cloud/HISTORY.md
@@ -2,7 +2,8 @@ Release History
 ===============
 3.0.1
 ---
-* Fix the deploy of Gradle project with multiple modules.
+* New preview argument `az spring-cloud app deploy` has new argument "--build-env" to specify build module and jvm version and so on.
+* Add error when `az spring-cloud app deploy` setting "--target-modules" and "--runtime-version for enterprise tier.
 
 3.0.0
 ---